refactor(ui): combine onboarding and support page (#4175)
