+++ ui/src/App.tsx
@@ -25,7 +25,6 @@ const Segments = loadable(() => import('./app/segments/Segments'));
 const Console = loadable(() => import('./app/console/Console'));
 const Login = loadable(() => import('./app/auth/Login'));
 const Settings = loadable(() => import('./app/Settings'));
-const Onboarding = loadable(() => import('./app/Onboarding'));
 const Support = loadable(() => import('./app/Support'));
 const Preferences = loadable(() => import('./app/preferences/Preferences'));
 const Namespaces = loadable(() => import('./app/namespaces/Namespaces'));
@@ -39,7 +38,7 @@ if (typeof window !== 'undefined') {
 
 const namespacedRoutes = [
   {
-    element: <Onboarding firstTime={true} />,
+    element: <Support firstTime={true} />,
     loader: () => {
       const state = store.getState();
       if (state?.user.completedOnboarding) {
@@ -116,10 +115,6 @@ const router = createHashRouter([
             element: <Namespaces />
           }
         ]
-      },
-      {
-        path: 'onboarding',
-        element: <Onboarding />
       },
       {
         path: 'support',
